requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
qrcode[pil]>=7.0.0
